+++ internal/data/time_series_model.go
@@ -1,4 +1,0 @@
-package data
-
-type TimeSeriesData struct {
-}